@@ -4,7 +4,6 @@ from __future__ import print_function
 import copy
 import os
 import shutil
-import glob
 import warnings
 import sys
 import importlib
@@ -92,7 +91,7 @@ jl_base_version = {
     "DashBase": "0.1",
 }
 
-jl_component_include_string = 'include("{name}.jl")'
+jl_component_include_string = 'include("jl/{name}.jl")'
 
 jl_resource_tuple_string = """DashBase.Resource(
     relative_package_path = {relative_package_path},
@@ -173,7 +172,7 @@ def get_jl_prop_types(type_object):
             )
         ),
         # React's PropTypes.objectOf
-        objectOf=lambda: ("Dict with Strings as keys and values of type {}").format(
+        objectOf=lambda: "Dict with Strings as keys and values of type {}".format(
             get_jl_type(type_object["value"])
         ),
         # React's PropTypes.shape
@@ -266,7 +265,7 @@ def create_docstring_jl(component_name, props, description):
     # Ensure props are ordered with children first
     props = reorder_props(props=props)
 
-    return ("A{n} {name} component.\n{description}\nKeyword arguments:\n{args}").format(
+    return "A{n} {name} component.\n{description}\nKeyword arguments:\n{args}".format(
         n="n" if component_name[0].lower() in "aeiou" else "",
         name=component_name,
         description=description,
@@ -284,7 +283,11 @@ def create_docstring_jl(component_name, props, description):
 
 
 def create_prop_docstring_jl(
-    prop_name, type_object, required, description, indent_num,
+    prop_name,
+    type_object,
+    required,
+    description,
+    indent_num,
 ):
     """
     Create the Dash component prop docstring
@@ -501,7 +504,13 @@ def generate_struct_file(name, props, description, project_shortname, prefix):
 
     file_name = format_fn_name(prefix, name) + ".jl"
 
-    file_path = os.path.join("src", file_name)
+    # put component files in src/jl subdir,
+    # this also creates the Julia source directory for the package
+    # if it is missing
+    if not os.path.exists("src/jl"):
+        os.makedirs("src/jl")
+
+    file_path = os.path.join("src", "jl", file_name)
     with open(file_path, "w") as f:
         f.write(import_string)
         f.write(class_string)
@@ -513,12 +522,7 @@ def generate_struct_file(name, props, description, project_shortname, prefix):
 def generate_module(
     project_shortname, components, metadata, pkg_data, prefix, **kwargs
 ):
-    # the Julia source directory for the package won't exist on first call
-    # create the Julia directory if it is missing
-    if not os.path.exists("src"):
-        os.makedirs("src")
-
-    # now copy over all JS dependencies from the (Python) components dir
+    # copy over all JS dependencies from the (Python) components dir
     # the inst/lib directory for the package won't exist on first call
     # create this directory if it is missing
     if os.path.exists("deps"):
@@ -526,14 +530,21 @@ def generate_module(
 
     os.makedirs("deps")
 
-    for javascript in glob.glob("{}/*.js".format(project_shortname)):
-        shutil.copy(javascript, "deps/")
+    for rel_dirname, _, filenames in os.walk(project_shortname):
+        for filename in filenames:
+            extension = os.path.splitext(filename)[1]
 
-    for css in glob.glob("{}/*.css".format(project_shortname)):
-        shutil.copy(css, "deps/")
+            if extension in [".py", ".pyc", ".json"]:
+                continue
 
-    for sourcemap in glob.glob("{}/*.map".format(project_shortname)):
-        shutil.copy(sourcemap, "deps/")
+            target_dirname = os.path.join(
+                "deps/", os.path.relpath(rel_dirname, project_shortname)
+            )
+
+            if not os.path.exists(target_dirname):
+                os.makedirs(target_dirname)
+
+            shutil.copy(os.path.join(rel_dirname, filename), target_dirname)
 
     generate_package_file(project_shortname, components, pkg_data, prefix)
     generate_toml_file(project_shortname, pkg_data)
