@@ -1,3 +1,4 @@
+# flake8: ignore=E501
 import json
 import os
 from collections import OrderedDict
@@ -27,17 +28,18 @@ expected_arg_strings = OrderedDict(
             "optionalSignature(shape)",
             "\n".join(
                 [
-                    "dict containing keys 'checked', 'children', 'customData', 'disabled', 'label', 'primaryText', 'secondaryText', 'style', 'value'.",
-                    "Those keys have the following types:",
-                    "- checked (boolean; optional)",
-                    "- children (a list of or a singular dash component, string or number; optional)",
-                    "- customData (bool | number | str | dict | list; required): A test description",
-                    "- disabled (boolean; optional)",
-                    "- label (string; optional)",
-                    "- primaryText (string; required): Another test description",
-                    "- secondaryText (string; optional)",
-                    "- style (dict; optional)",
-                    "- value (bool | number | str | dict | list; required)",
+                    "dict with keys:\n",
+                    "    - checked (boolean; optional)\n",
+                    "    - children (a list of or a singular dash component, string or number; optional)\n",
+                    "    - customData (bool | number | str | dict | list; required):\n"
+                    "        A test description.\n",
+                    "    - disabled (boolean; optional)\n",
+                    "    - label (string; optional)\n",
+                    "    - primaryText (string; required):\n"
+                    "        Another test description.\n",
+                    "    - secondaryText (string; optional)\n",
+                    "    - style (dict; optional)\n",
+                    "    - value (bool | number | str | dict | list; required)",
                 ]
             ),
         ],
@@ -45,20 +47,19 @@ expected_arg_strings = OrderedDict(
             "requiredNested",
             "\n".join(
                 [
-                    "dict containing keys 'customData', 'value'.",
-                    "Those keys have the following types:",
-                    "- customData (dict; required): customData has the following type: dict containing keys 'checked', 'children', 'customData', 'disabled', 'label', 'primaryText', 'secondaryText', 'style', 'value'.",
-                    "  Those keys have the following types:",
-                    "  - checked (boolean; optional)",
-                    "  - children (a list of or a singular dash component, string or number; optional)",
-                    "  - customData (bool | number | str | dict | list; required)",
-                    "  - disabled (boolean; optional)",
-                    "  - label (string; optional)",
-                    "  - primaryText (string; required)",
-                    "  - secondaryText (string; optional)",
-                    "  - style (dict; optional)",
-                    "  - value (bool | number | str | dict | list; required)",
-                    "- value (bool | number | str | dict | list; required)",
+                    "dict with keys:\n",
+                    "    - customData (dict; required)\n\n"
+                    "        `customData` is a dict with keys:\n",
+                    "        - checked (boolean; optional)\n",
+                    "        - children (a list of or a singular dash component, string or number; optional)\n",
+                    "        - customData (bool | number | str | dict | list; required)\n",
+                    "        - disabled (boolean; optional)\n",
+                    "        - label (string; optional)\n",
+                    "        - primaryText (string; required)\n",
+                    "        - secondaryText (string; optional)\n",
+                    "        - style (dict; optional)\n",
+                    "        - value (bool | number | str | dict | list; required)\n",
+                    "    - value (bool | number | str | dict | list; required)",
                 ]
             ),
         ],
@@ -71,46 +72,78 @@ expected_doc = [
     "It's multiple lines long.",
     "",
     "Keyword arguments:",
-    "- requiredString (string; required): A required string",
-    "- optionalString (string; default ''): A string that isn't required.",
-    "- optionalBoolean (boolean; default False): A boolean test",
-    "- optionalNode (a list of or a singular dash component, string or number; optional): "
-    "A node test",
-    "- optionalArray (list; optional): An array test with a particularly ",
-    "long description that covers several lines. It includes the newline character ",
-    "and should span 3 lines in total.",
-    "- requiredUnion (string | number; required)",
-    "- optionalSignature(shape) (dict; optional): This is a test of an object's shape. "
-    "optionalSignature(shape) has the following type: dict containing keys 'checked', "
-    "'children', 'customData', 'disabled', 'label', 'primaryText', 'secondaryText', "
-    "'style', 'value'.",
-    "  Those keys have the following types:",
-    "  - checked (boolean; optional)",
-    "  - children (a list of or a singular dash component, string or number; optional)",
-    "  - customData (bool | number | str | dict | list; required): A test description",
-    "  - disabled (boolean; optional)",
-    "  - label (string; optional)",
-    "  - primaryText (string; required): Another test description",
-    "  - secondaryText (string; optional)",
-    "  - style (dict; optional)",
-    "  - value (bool | number | str | dict | list; required)",
-    "- requiredNested (dict; required): requiredNested has the following type: dict containing "
-    "keys 'customData', 'value'.",
-    "  Those keys have the following types:",
-    "  - customData (dict; required): customData has the following type: dict containing "
-    "keys 'checked', 'children', 'customData', 'disabled', 'label', 'primaryText', "
-    "'secondaryText', 'style', 'value'.",
-    "    Those keys have the following types:",
+    "",
+    "- optionalArray (list; optional):",
+    "    An array test with a particularly  long description that covers",
+    "    several lines. It includes the newline character  and should span",
+    "    3 lines in total.",
+    "",
+    "- optionalBoolean (boolean; default False):",
+    "    A boolean test.",
+    "",
+    "- optionalNode (a list of or a singular dash component, string or number; optional):",
+    "    A node test.",
+    "",
+    "- optionalSignature(shape) (dict; optional):",
+    "    This is a test of an object's shape.",
+    "",
+    "    `optionalSignature(shape)` is a dict with keys:",
+    "",
     "    - checked (boolean; optional)",
+    "",
     "    - children (a list of or a singular dash component, string or number; optional)",
-    "    - customData (bool | number | str | dict | list; required)",
+    "",
+    "    - customData (bool | number | str | dict | list; required):",
+    "        A test description.",
+    "",
     "    - disabled (boolean; optional)",
+    "",
     "    - label (string; optional)",
-    "    - primaryText (string; required)",
+    "",
+    "    - primaryText (string; required):",
+    "        Another test description.",
+    "",
     "    - secondaryText (string; optional)",
+    "",
     "    - style (dict; optional)",
+    "",
     "    - value (bool | number | str | dict | list; required)",
-    "  - value (bool | number | str | dict | list; required)",
+    "",
+    "- optionalString (string; default ''):",
+    "    A string that isn't required.",
+    "",
+    "- requiredNested (dict; required)",
+    "",
+    "    `requiredNested` is a dict with keys:",
+    "",
+    "    - customData (dict; required)",
+    "",
+    "        `customData` is a dict with keys:",
+    "",
+    "        - checked (boolean; optional)",
+    "",
+    "        - children (a list of or a singular dash component, string or number; optional)",
+    "",
+    "        - customData (bool | number | str | dict | list; required)",
+    "",
+    "        - disabled (boolean; optional)",
+    "",
+    "        - label (string; optional)",
+    "",
+    "        - primaryText (string; required)",
+    "",
+    "        - secondaryText (string; optional)",
+    "",
+    "        - style (dict; optional)",
+    "",
+    "        - value (bool | number | str | dict | list; required)",
+    "",
+    "    - value (bool | number | str | dict | list; required)",
+    "",
+    "- requiredString (string; required):",
+    "    A required string.",
+    "",
+    "- requiredUnion (string | number; required)",
 ]
 
 
@@ -129,13 +162,13 @@ def test_docstring(load_test_flow_metadata_json):
         load_test_flow_metadata_json["props"],
         load_test_flow_metadata_json["description"],
     )
+    print(docstring.splitlines())
     prohibit_events(load_test_flow_metadata_json["props"]),
     assert not list(unified_diff(expected_doc, docstring.splitlines()))
 
 
 def test_docgen_to_python_args(load_test_flow_metadata_json):
     props = load_test_flow_metadata_json["props"]
-
     for prop_name, prop in list(props.items()):
         assert (
             js_to_py_type(prop["flowType"], is_flow_type=True)
